% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVEM.R
\name{AVEM}
\alias{AVEM}
\title{Array Vector-Element Multiplication}
\usage{
AVEM(A, v)
}
\arguments{
\item{A}{two or three-dimensional array.}

\item{v}{vector}
}
\value{
array resulting from multiplying the matrices/columns in A by the corresponding elements in v.
}
\description{
Multiply each matrix/column in a 3D/2D array by the corresponding element of a vector.
}
\examples{
A <- array( 1:24, dim = c( 2,3,4 ) )
b = 1:4
AVEM( A, b )
}
