% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLA_2perpB.r
\name{BLA_2perpB}
\alias{BLA_2perpB}
\title{Bayes Linear Adjustment by 2 Perpendicular Known Boundaries}
\usage{
BLA_2perpB(
  x,
  K_d,
  L_d,
  xK = NA,
  xL = NA,
  xLK = NA,
  xK_d = NA,
  xL_d = NA,
  fxK,
  fxL,
  fxLK,
  E_fx = 0,
  E_fxK = 0,
  E_fxL = 0,
  E_fxLK = 0,
  theta,
  s2
)
}
\arguments{
\item{x}{points at which we want to update}

\item{K_d}{the dimensions which, when fixed at certain values, result in known boundary K.}

\item{L_d}{the dimensions which, when fixed at certain values, result in known boundary L.}

\item{xK}{the projection of x onto known boundary K}

\item{xL}{the projection of x onto known boundary L}

\item{xLK}{the projection of x onto the intersection of known boundaries K and L.}

\item{xK_d}{values the dimensions K must take for the function to be known}

\item{xL_d}{values the dimensions L must take for the function to be known}

\item{fxK}{function evaluated at x projected onto the boundary K.}

\item{fxL}{function evaluated at x projected onto the boundary L.}

\item{fxLK}{function evaluated at x projected onto the intersection of boundaries K and L.}

\item{E_fx}{prior expectation for the function f(x)}

\item{E_fxK}{prior expectation for f(x^K)}

\item{E_fxL}{prior expectation for f(x^L)}

\item{E_fxLK}{prior expectation for f(x^LK)}

\item{theta}{vector of correlation length parameter values.}

\item{s2}{scalar variance parameter value.}
}
\value{
\item{EB_fx}{Expected value of f(x) adjusted by knowledge of function behaviour along K and L.}
\item{VarB_fx}{Variance of f(x) adjusted by knowledge of function behaviour along K and L.}
\item{CovB_fx}{Covariance of f(x) adjusted by knowledge of function behaviour along K and L.}
}
\description{
Perform a Bayes linear adjustment utilising knowledge of function behaviour along
two perpendicular known boundaries in the input space.
}
\examples{
# Toy function
f <- function( x ){

sin( x[1] / ( exp( x[2] ) ) ) + cos( x[3] )

}
x <- matrix( runif( 12 ), ncol = 3 )
K_d = 2
L_d = 1
xK_d = 0
xL_d = 0
fxK <- f_boundary( x = x, K_d = K_d, xK_d = xK_d, f = f )
fxL <- f_boundary( x = x, K_d = L_d, xK_d = xL_d, f = f )
fxLK <- f_boundary( x = x, K_d = c(K_d, L_d), xK_d = c(xK_d, xL_d), f = f )
theta <- c( pi, pi/8, pi )
s2 <- 2
BA <- BLA_2perpB( x = x, K_d = K_d, L_d = L_d, xK_d = xK_d, xL_d = xL_d,
                  fxK = fxK, fxL = fxL, fxLK = fxLK, theta = theta, s2 = s2 )
}
