% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_for_plot.r
\name{boundary_for_plot}
\alias{boundary_for_plot}
\title{Boundary Coordinate Generation}
\usage{
boundary_for_plot(fixed_dimension, fixed_value, ranges)
}
\arguments{
\item{fixed_dimension}{the dimensions which are fixed for the boundary}

\item{fixed_value}{the values for the coordinates which are fixed (of same length as fixed_dimension).}

\item{ranges}{ranges of the three variables, given as a 3 x 2 matrix.}
}
\value{
A matrix with 2 columns, and 4 rows for 2D boundaries and 2 rows for 1D boundaries
This gives the coordinates of the boundaries for plotting on the cube as shown in the top
row of the 3D example figures in the article.
}
\description{
A function for generating the coordinate matrices
(with 2 columns, and 4 rows for 2D boundaries and 2 rows for 1D boundaries)
required for the cube.
}
\examples{
ranges <- matrix(c( -2*pi, 2*pi,
                    -pi/4, pi/4,
                    -2*pi, 2*pi), ncol = 2, byrow = TRUE)
boundary_for_plot( fixed_dimension = 2,
                   fixed_value = 0,
                   ranges = ranges )
}
