% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.r
\name{contour_plot}
\alias{contour_plot}
\title{Generate Contour Plot}
\usage{
contour_plot(x, y, z, levels, colours)
}
\arguments{
\item{x}{x coordinates.}

\item{y}{y coordinates.}

\item{z}{matrix of values to be plotted, with rows assumed to correspond to increasing values of x (from top to bottom)}

\item{levels}{levels at which z should be divided into for the contour plot.}

\item{colours}{colours for each level of z.}
}
\value{
nothing is returned.  Contour plot is generated.
}
\description{
User-friendly wrapper for generating the contour plots as of the 3D example shown in the text.
}
\examples{
x = seq( from = 0 , to = 1, by = 0.01 )
y = seq( from = 0 , to = 1, by = 0.01 )
eg <- expand.grid( x, y )
z <- matrix( apply( eg, 1, FUN = function(x)( cos(x[1]) + tan(x[2]) )^2 ), nrow = 101 )
nlevels = 20
levels <- pretty( c(0,7), nlevels )
colours <- grDevices::colorRampPalette(
               c( "white", "cyan", "blue", "purple", "pink" ), space = "Lab" )
contour_plot( x = x, y = y, z = z, levels = levels, colours = colours )
}
