% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cube.r
\name{draw_cube}
\alias{draw_cube}
\title{Draw a Cube}
\usage{
draw_cube(
  lwd = 2,
  lty2 = 2,
  cex = 1.8,
  col_line_width = lwd,
  coloured_hp = list(),
  col_hp = c("green", "red", "blue", "pink"),
  density_col = rep(0.7, length(coloured_hp)),
  main = "",
  cex.main = 1,
  main.line = 1
)
}
\arguments{
\item{lwd}{line width for the edges of the cube.}

\item{lty2}{line type for the "hidden" edges of the cube.}

\item{cex}{scale size for the labels.}

\item{col_line_width}{line width for the coloured lines representing 1D boundaries.}

\item{coloured_hp}{a list of vectors, where each vector represents the hyperplanes to be plotted:}

\item{col_hp}{the colour of the hyperplanes given above.}

\item{density_col}{the density of the fill of the 2D hyperplanes (note that the length of this vector}

\item{main}{title of the cube plot.}

\item{cex.main}{size of the title.}

\item{main.line}{number of lines outwards from the plot edge to plot the title.}
}
\value{
nothing is returned.  Cube is plotted.
}
\description{
Specific function for the diagrams of a cube illustrating which boundaries are known and
which cross-section of the input space is being emulated.
}
\examples{
ranges <- matrix(c( -2*pi, 2*pi,
                    -pi/4, pi/4,
                    -2*pi, 2*pi), ncol = 2, byrow = TRUE)
boundary_for_plot <- boundary_for_plot( fixed_dimension = 2,
                                        fixed_value = 0,
                                        ranges = ranges )
draw_cube( coloured_hp = list( boundary_for_plot ) )
}
