% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_boundary.R
\name{f_boundary}
\alias{f_boundary}
\title{Evaluate f along a boundary.}
\usage{
f_boundary(x, K_d, xK_d, f)
}
\arguments{
\item{x}{set of points, given as a matrix.}

\item{K_d}{Variable indices to be fixed.}

\item{xK_d}{Values at which those variables are fixed to.}

\item{f}{a toy function which is to be evaluated.}
}
\value{
Value of f(x) at the projections of x projected onto boundary K_d.
}
\description{
Evaluate f along a boundary.
}
\examples{
# Toy function
f <- function( x ){

sin( x[1] / ( exp( x[2] ) ) ) + cos( x[3] )

}
x <- matrix( runif( 12 ), ncol = 3 )
f_boundary( x = x,
            K_d = 2,
            xK_d = 0,
            f = f )
}
