% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_generation.r
\name{legend_generation}
\alias{legend_generation}
\title{Legend Generation}
\usage{
legend_generation(colours, levels)
}
\arguments{
\item{colours}{colours for each level of z.}

\item{levels}{levels at which z should be divided into for the contour plot.}
}
\value{
nothing is returned.  Legend is plotted.
}
\description{
Generate the legend for the 3D example as shown in the article.
}
\examples{
nlevels = 20
levels <- pretty( c(0,7), nlevels )
colours <- grDevices::colorRampPalette(
               c( "white", "cyan", "blue", "purple", "pink" ), space = "Lab" )
legend_generation( colours = colours, levels = levels )
}
