% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.r
\name{plot_setup}
\alias{plot_setup}
\title{Plot Setup for Article.}
\usage{
plot_setup()
}
\value{
Sets up plot domain.
}
\description{
A simple function for setting up the plot domain to generate figures similar to those
presented for the 3D example shown in the article.
}
\examples{
plot_setup()
}
