% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGen3dEx.r
\name{PlotGen3dEx}
\alias{PlotGen3dEx}
\title{3D Example Plot Generation}
\usage{
PlotGen3dEx(
  f,
  ranges,
  K_d,
  L_d = NA,
  M_d = NA,
  xK_d,
  xL_d = NA,
  xM_d = NA,
  fixed_dimension,
  fixed_value,
  theta = rep(pi/2, 3),
  s2 = 1,
  grid_length = 50,
  lwd = 2,
  lty2 = 2,
  cex = 1.8,
  col_line_width = 3,
  zlim_f = "assessed",
  zlim_var = "assessed",
  zlim_diag = c(-4.25, 4.25),
  legend = FALSE,
  main_cube = "",
  cex.main_cube = 1.8,
  main.line = 0.2
)
}
\arguments{
\item{f}{a toy function for which plots will be created.}

\item{ranges}{the ranges for the input parameters of the toy functions, given as a matrix.}

\item{K_d}{variable indices to be fixed for boundary K.}

\item{L_d}{variable indices to be fixed for boundary L.}

\item{M_d}{variable indices to be fixed for boundary M.}

\item{xK_d}{values at which those variables are fixed to.}

\item{xL_d}{values at which those variables are fixed to.}

\item{xM_d}{values at which those variables are fixed to.}

\item{fixed_dimension}{index of the variable that will be kept fixed for the plots.}

\item{fixed_value}{fixed value in the remaining dimension.}

\item{theta}{correlation length parameters.}

\item{s2}{scalar variance parameter.}

\item{grid_length}{number of grid points along each dimension with which to represent the plotted surface.}

\item{lwd}{line width for cube edges.}

\item{lty2}{line type for background lines}

\item{cex}{relative size of plot}

\item{col_line_width}{width of coloured boundary lines}

\item{zlim_f}{plotting range for the z-values (model and mean prediction)}

\item{zlim_var}{plotting range for the variance predictions.}

\item{zlim_diag}{plotting range for the diagnostic plot.}

\item{legend}{Add a legend to the side of the plot?}

\item{main_cube}{title for the cube plot.}

\item{cex.main_cube}{font size of the cube plot title.}

\item{main.line}{line value for the title function.}
}
\value{
nothing is returned.  Plots are generated.
}
\description{
Generic Generation of Plots for 3D Toy Example.
}
\examples{
# Specify the toy function - requires a 3D input vector.
f <- function( x ){

 sin( x[1] / ( exp( x[2] ) ) ) + cos( x[3] )

}

# Specify ranges for the function parameters x1, x2, and x3.
ranges <- matrix(c( -2*pi, 2*pi,
                    -pi/4, pi/4,
                    -2*pi, 2*pi), ncol = 2, byrow = TRUE)

# Specify the correlation length parameters and variance parameter for the example.
theta <- c( pi, pi/8, pi )
s2 <- 2

# Specify the ranges for the function/mean and variance plots.
zlim_f <- c(-2.5, 2.5)
zlim_var <- c(0, 2)

plot_setup()
PlotGen3dEx( f = f,
             ranges = ranges,
             K_d = c(2,3),
             L_d = c(2,3),
             M_d = 1,
             xK_d = c(0,0),
             xL_d = c(0,-pi),
             xM_d = 0,
             fixed_dimension = 2,
             fixed_value = 0,
             theta = theta,
             s2 = s2,
             zlim_f = zlim_f,
             zlim_var = zlim_var,
             main_cube = bquote( x[2] == 0 ),
             legend = FALSE )
}
